pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.2.1",
]
